Check T(O) in 4

+++ DataProtection/DataProtection/PageModels/Lab1/NumbersToPower.cs
@@ -8,8 +8,8 @@ namespace DataProtection.PageModels.Lab1
     public class NumbersToPower
     {
         public long a { get; set; }
-        public int x { get; set; }
+        public long x { get; set; }
         public long y { get; set; }
-        public int p { get; set; }
+        public long p { get; set; }
     }
 }
